# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zed CV API
  description: AI-powered job matching platform for Zambia
  version: 0.1.0
  contact:
    email: convergeozambia@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.zedcv.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OTPRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string, pattern: '^\+260[0-9]{9}$', example: "+260971234567" }
    OTPVerify:
      type: object
      required: [phone, code]
      properties:
        phone: { type: string, pattern: '^\+260[0-9]{9}$' }
        code: { type: string, minLength: 6, maxLength: 6 }
    AuthTokens:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        user_id: { type: string, format: uuid }
...
    AdminStats:
      type: object
      properties:
        users_total: { type: integer }
        users_active_30d: { type: integer }
        subscriptions_active: { type: integer }
        subscriptions_paid: { type: integer }
        jobs_total: { type: integer }
        jobs_active: { type: integer }
        jobs_expired: { type: integer }
        matches_24h: { type: integer }
        matches_total: { type: integer }
        revenue_ngwee_30d: { type: integer }
        revenue_ngwee_total: { type: integer }
    AdminUserRow:
      type: object
      properties:
        id: { type: string, format: uuid }
        phone: { type: string }
        full_name: { type: string }
        location: { type: string }
        subscription_tier: { type: string, enum: [free, starter, professional] }
        role: { type: string, enum: [user, superadmin] }
        matches_used: { type: integer }
        matches_limit: { type: integer }
        created_at: { type: string, format: date-time }
    AdminUserList:
      type: object
      properties:
        users: { type: array, items: { $ref: '#/components/schemas/AdminUserRow' } }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        pages: { type: integer }
    AdminJobRow:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        source: { type: string, enum: [manual, scraper, ocr, partner] }
        quality_score: { type: integer }
        is_active: { type: boolean }
        closing_date: { type: string, format: date }
        posted_at: { type: string, format: date-time }
    AdminJobList:
      type: object
      properties:
        jobs: { type: array, items: { $ref: '#/components/schemas/AdminJobRow' } }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        pages: { type: integer }
    BulkDeactivateRequest:
      type: object
      properties:
        job_ids: { type: array, items: { type: string, format: uuid }, maxItems: 500 }
        expired_only: { type: boolean, default: false }
    BulkDeactivateResponse:
      type: object
      properties:
        deactivated: { type: integer }
    AdminPaymentRow:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        user_phone: { type: string }
        amount: { type: integer, description: "ngwee" }
        currency: { type: string }
        payment_method: { type: string }
        provider: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
    AdminPaymentList:
      type: object
      properties:
        payments: { type: array, items: { $ref: '#/components/schemas/AdminPaymentRow' } }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        pages: { type: integer }
        total_completed_ngwee: { type: integer }
...
    Job:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description: { type: string }
        requirements: { type: array, items: { type: string } }
        skills_required: { type: array, items: { type: string } }
        salary_min: { type: integer }
        salary_max: { type: integer }
        apply_url: { type: string, format: uri }
        apply_email: { type: string, format: email }
        source: { type: string, enum: [manual, scraper, ocr, partner] }
        quality_score: { type: integer, minimum: 0, maximum: 100 }
        closing_date: { type: string, format: date }
        posted_at: { type: string, format: date-time }
        is_active: { type: boolean }
    JobCreate:
      type: object
      required: [title, description, source]
      properties:
        title: { type: string, minLength: 5 }
        company: { type: string }
        location: { type: string }
        description: { type: string, minLength: 20 }
        requirements: { type: array, items: { type: string } }
        skills_required: { type: array, items: { type: string } }
        salary_min: { type: integer }
        salary_max: { type: integer }
        apply_url: { type: string, format: uri }
        apply_email: { type: string, format: email }
        source: { type: string, enum: [manual, scraper, ocr, partner] }
        closing_date: { type: string, format: date }
    JobList:
      type: object
      properties:
        jobs: { type: array, items: { $ref: '#/components/schemas/Job' } }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
...
paths:
...
  /auth/otp/request:
    post:
      operationId: requestOTP
      summary: Request OTP via WhatsApp
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OTPRequest' }
      responses:
        '200': { description: OTP sent }
        '429': { description: Rate limited }
  /auth/otp/verify:
    post:
      operationId: verifyOTP
      summary: Verify OTP and get tokens
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OTPVerify' }
      responses:
        '200':
          description: Auth successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401': { description: Invalid OTP }
...
  /admin/stats:
    get:
      operationId: adminStats
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminStats' }
        '403': { description: Superadmin only }
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: tier, in: query, schema: { type: string, enum: [free, starter, professional] } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUserList' }
        '403': { description: Superadmin only }
  /admin/jobs:
    get:
      operationId: adminListJobs
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25 } }
        - { name: expired, in: query, schema: { type: boolean } }
        - { name: is_active, in: query, schema: { type: boolean } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminJobList' }
        '403': { description: Superadmin only }
  /admin/jobs/bulk-deactivate:
    post:
      operationId: adminBulkDeactivateJobs
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkDeactivateRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulkDeactivateResponse' }
        '403': { description: Superadmin only }
  /admin/payments:
    get:
      operationId: adminListPayments
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25 } }
        - { name: status, in: query, schema: { type: string, enum: [pending, completed, failed, refunded] } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPaymentList' }
        '403': { description: Superadmin only }
...
  /jobs:
    get:
      operationId: listJobs
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 20 } }
        - { name: location, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobList' }
    post:
      operationId: createJob
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
  /jobs/{job_id}:
    get:
      operationId: getJob
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: job_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
...